feat(solve_time): extend to seconds with precise timestamps
- Change solve_time from u32 (days) to u64 (seconds)
- Update date format from YYYY-MM-DD to YYYY-MM-DD HH:MM:SS
- Add format_duration_human_readable() for display (Xy Xmo Xd Xh Xm)
- Fetch timestamps from mempool.space API (claim tx time)
- Update 82 b1000 puzzles and 2 hash_collision puzzles
- Update tests to validate seconds-based solve_time

Closes #27

+++ build.rs
@@ -38,7 +38,7 @@ struct Btc1000Puzzle {
     pubkey_format: Option<String>,
     start_date: Option<String>,
     solve_date: Option<String>,
-    solve_time: Option<u32>,
+    solve_time: Option<u64>,
     source_url: Option<String>,
 }
 
@@ -63,7 +63,7 @@ struct HashCollisionPuzzle {
     prize: Option<f64>,
     start_date: Option<String>,
     solve_date: Option<String>,
-    solve_time: Option<u32>,
+    solve_time: Option<u64>,
     source_url: Option<String>,
 }
 
@@ -88,7 +88,7 @@ struct GsmgPuzzle {
     pubkey_format: Option<String>,
     start_date: Option<String>,
     solve_date: Option<String>,
-    solve_time: Option<u32>,
+    solve_time: Option<u64>,
     source_url: Option<String>,
 }
